feat(app): adicionada página home e paginas expense e expense detail

+++ Budget.App/Models/BudgetValue.cs
@@ -5,8 +5,8 @@ namespace Budget.App.Models;
 public class BudgetValue : NotifyPropertyChanged
 {
     private string _categoria = null!;
-    private double _valor;
+    private decimal _valor;
     public string Categoria { get => _categoria; set => SetField(ref _categoria, value); }
-    public double Valor { get => _valor; set => SetField(ref _valor, value); }
+    public decimal Valor { get => _valor; set => SetField(ref _valor, value); }
 
 }
